# Proxy/DOC/UnifiedAPIProxy-OpenAPI-1.0.0.yaml
openapi: 3.0.3
info:
  title: Unified API Proxy
  description: |
    Unified proxy service for accessing Also Marketplace, TDSynnex StreamOne, and Trend Vision One APIs through a single gateway.
    
    ## Features
    - **Single Entry Point**: One proxy for three different APIs
    - **Automatic Request Forwarding**: Transparent proxy with header and body forwarding
    - **Centralized Logging**: Winston-based logging for all requests
    - **Health Monitoring**: Built-in health checks and status endpoints
    
    ## Architecture
    ```
    Client Application → Unified Proxy (Port 4000) → Backend APIs
                         ├─ /api/also/*      → Also API (Port 3001)
                         ├─ /api/tdsynnex/*  → TDSynnex API (Port 3002)
                         └─ /api/trend/*     → Trend API (Port 3003)
    ```
    
    ## Authentication
    The proxy transparently forwards authentication headers to backend APIs:
    - **Also**: `Authenticate: CCPSessionId <token>`
    - **TDSynnex**: `Authorization: Bearer <token>`
    - **Trend**: `Authorization: Bearer <api_key>`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@company.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.company.com
    description: Production server
...
paths:
...
  /api/also/{path}:
    x-summary: Also Marketplace proxy
    parameters:
      - name: path
        in: path
        required: true
        description: The path to forward to Also Marketplace API
        schema:
          type: string
        example: auth/login
      - name: Authenticate
        in: header
        description: Also session token (format CCPSessionId <token>)
        schema:
          type: string
        example: CCPSessionId C9C5MlqrpMwG/6bQ3mAVlm0Z6hi/eh1vsQs4i1I/g==
    get:
      summary: Proxy GET request to Also API
      description: Forwards GET request to Also Marketplace API at http://localhost:3001/api/{path}
      tags:
        - Also Marketplace
      responses:
        '200':
          description: Successful response from Also API
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
    post:
      summary: Proxy POST request to Also API
      description: Forwards POST request to Also Marketplace API
      tags:
        - Also Marketplace
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Successful response from Also API
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Proxy PATCH request to Also API
      description: Forwards PATCH request to Also Marketplace API
      tags:
        - Also Marketplace
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Successful response from Also API
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Proxy DELETE request to Also API
      description: Forwards DELETE request to Also Marketplace API
      tags:
        - Also Marketplace
      responses:
        '200':
          description: Successful response from Also API
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
    UnauthorizedError:
      description: Authentication failed or missing credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Invalid or missing authentication token
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Route not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Internal server error
    ServiceUnavailableError:
      description: Backend service is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: connect ECONNREFUSED 127.0.0.1:3001
...
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: object
          description: Additional error details
        code:
          type: string
          description: Error code
      required:
        - success
        - message
  securitySchemes:
    AlsoSessionToken:
      type: apiKey
      in: header
      name: Authenticate
      description: |
        Also Marketplace session token authentication.
        Format: `CCPSessionId <token>`
        
        Obtain token via `/api/also/auth/login`

    OAuth2Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        TDSynnex OAuth2 Bearer token authentication.
        
        Obtain token via `/api/tdsynnex/auth/token`

    TrendApiKey:
      type: http
      scheme: bearer
      description: |
        Trend Vision One API token authentication.
        
        Configure in Trend Vision One console
security:
  - AlsoSessionToken: []
  - OAuth2Bearer: []
  - TrendApiKey: []